Button Converter FIXED
Issue: clicking back and forth between C and F was working, and because I added temp = null, it wasn't double covnerting C, but, when you left it on F and searched for a new city, the temp nubmersstill always get converted to C, but the button was displaying as F being active. It was as minor bug because when you tried to click F there still weren't any chasnges in numbers, and when you clicked C it simply changed the active id as C being active, but still wanted to fix it. this version has that fixed, added a statement to the submit button that forces the active id to celcius every time search button is clicked